pyproject.toml
[project]
name = "snipster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Oh-Really", email = "moreilly406@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[all]>=0.121.2",
    "psycopg2>=2.9.11",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.27",
    "typer>=0.20.0",
]

[project.scripts]
#snipster = "snipster:main"
snipster = "snipster.cli:app"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.11",
    "pytest-cov>=7.0.0",
]
